+++ lib/generator.py
@@ -133,45 +133,3 @@ class GridGenerator():
         self.grid = grid_manager.grid
         self.replace_none_start()
         grid_manager.grid = self.grid
-
-
-if __name__ == "__main__":
-    # TODO: WRITE THIS AS A TEST
-    
-    def check_any_adjacence(grid):
-        """
-            Returns true if the grid is invalid
-        """
-
-        res = True
-
-        for (offseted_line_id, line_slice) in enumerate(grid[1:-1]):
-            line_id = offseted_line_id - 1
-            for (offseted_element_position, element) in enumerate(line_slice[1:-1]):
-                element_position = offseted_element_position - 1
-
-                res = res and not ((element == grid[line_id-1][element_position] and \
-                        element == grid[line_id+1][element_position]) or \
-                        (element == grid[line_id][element_position-1] and \
-                        element == grid[line_id][element_position+1]))
-
-        return res
-
-    N_TEST = 1000
-    print(f"Testing for {N_TEST} random tests")
-
-    i = 0
-    ONE_FAILED = False
-    while i <= N_TEST and not ONE_FAILED:
-        grid_instance = GridGenerator()
-
-        r_grid = grid_instance.init_sequence(randint(10, 30), randint(10, 30), randint(4, 5))
-
-        c_res = check_any_adjacence(r_grid)
-
-        if not c_res:
-            print("  Test passed")
-            i += 1
-        else:
-            print("  Test failed")
-            ONE_FAILED = True
